requirements.txt
numpy
groq
asgiref==3.9.1
beautifulsoup4==4.13.5
Django==5.2.5
dnspython==2.7.0
faiss-cpu==1.12.0
pydantic
pymongo==4.14.1
python-dotenv
selenium==4.35.0
sentence-transformers==5.1.0
urllib3==2.5.0
webdriver-manager==4.0.2
django-cors-headers
djangorestframework
djangorestframework-simplejwt